Move DeviceTypeTestDsl
