+++ bowler-kernel/hardware/src/test/kotlin/com/neuronrobotics/bowlerkernel/hardware/device/deviceid/Esp32WROOM32Test.kt
@@ -16,7 +16,6 @@
  */
 package com.neuronrobotics.bowlerkernel.hardware.device.deviceid
 
-import com.neuronrobotics.bowlerkernel.hardware.deviceTypeTest
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.Timeout
 import java.util.concurrent.TimeUnit
